+++ abci/app/v1/ndid.go
@@ -61,4 +61,6 @@ var regulatorMethod = map[string]bool{
 	"RemoveRequestType":                                    true,
 	"AddSuppressedIdentityModificationNotificationNode":    true,
 	"RemoveSuppressedIdentityModificationNotificationNode": true,
+	"AddAllowedNodeSupportedFeature":                       true,
+	"RemoveAllowedNodeSupportedFeature":                    true,
 }